requirements.txt
torch
python-dotenv
spotipy